Removed Excel output files from Test Models folder, so Standalone zip file is under the Github limit
Recompiled Pip installer files and standalone code. Need to test installers and run test models before ready for release

Modified FractureDipSet so it does not revert from ResidualActive to Growing within the calculateTotalMacrofracturePopulation() function, but instead within the fds.checkStressShadowWidthChange() function, only if the stress shadow width has changed. This fixes the problem (B75) but still needs testing on multi-stage models

A number of Petrel UI enhancements, including: 1) Automatically close any Deformation Episode UI dialogs when the DFM Generator dialog is closed using the Cancel button; 2) Give the user a warning if there are any Deformation Episode UI dialogs open when clicking Apply or OK on the DFM Generator dialog; 3) Not closing the DFM Generator dialog if it gives a warning rather than running the model when the OK button is clicked; 4) Not allowing the Deformation Episode UI dialog to be resized so that some controls are not visible; 5) Make sure "Add, Edit and Remove Deformation Episode" buttons on the DFM Generator dialog are anchored and do not move when the dilaog is resized (ME60). Also added the WorkToDo spreadsheet listing known (outstanding and solved) bugs and enhancements to the Git repository. Not yet recompiled

In Grid.CalculateAllFractureData(), put the calls to Gridblock.CalculateFractureData() in a try block so any errors in individual gridblocks can be picked up and error messages output, without aborting the entire model

Modified FractureDipSet so that mode factors are recalculated dynamically when required based on geometric factors (B83), and modified GridblockConfiguration.PropagateDFN() and DFNGenerationControl to apply a limit to the number of explicit fractures nucleating per timestep per gridblock, to prevent the code from hanging when excessive fractures nucleate (B83). Also added return codes to GridblockConfiguration.CalculateFractureData() and GridblockConfiguration.PropagateDFN() to give details of any errors, and modified Grid functions CalculateAllFractureData() and PropagateLocalDFNs() to output error messages based on these codes.

Updated errors in comments - no change to code

Recompiled Petrel installers and standalone executable